🚚 rename: separate community main component
